Functioning Casino Structure
Finally I am at a point in the project where I can add games to the project, to expand the casino. All framework items are done referring to certain options the user can choose to navigate through the casino. However more will be to come like... a bar with really expensive champagne for the rich to bathe in